parse rawData twice to make sure that we get the JSON object
consider the following string "{\"a\":2,\"b\":4,\"c\":6}". When JSON.parse is called on this string, we get "{"a":2,"b":4,"c":6}", but here is the catch: this is a string. So we have to run JSON.parse twice to get the actual object :clown_face:

This caused the JSON on the data browser to be shown as a string as opposed to a prettified JSON.

See this SO post for more details https://stackoverflow.com/questions/42494823/json-parse-returns-string-instead-of-object