call toast with error message on fetch error
databrowser simply shows SidebarMissingData when there is an error which complicates debugging the issues. This was an issue when we debugged the case when two integer keys were compared. It's also a problem in a recent ticket.

This change adds a toast when there is an error. This will show the error text to the user who can share it with us.